Add files via upload
ASSIGNMENT 1:
REPORT ON VISIT TO NALI 2023
Reflection:
During the visit, I visited many booths and concluded that technology is essential to us for improving the quality of life. Most of the applications and products that were invented are very useful and help us facilitate our working process and improve our quality of life. Beside that,I also gained many fascinating ideas from the interviewee after the interview session. However, the sound of NALI’s briefing is too loud, which causes the sounds of the interviewer and interviewee to not be clear during the interview session. Overall, I think these activities are crucial to the students, and I also have a whale of time at there!